Finished Up Character Controller v0.1
Changes made:
- Basic movement added
- Basic camera controls added
- Code cleaned up

Reading in saved settings for sensitivity and camera inversion will need addressed once UI and settings are created.

Special note: Bug was located in the previous version that may effect other areas in the future. Network Transform component was added to the player, the intent was to prevent de-syncing between clients. However, only the host experienced smooth and responsive movement. For now, all players have had network transform removed and a custom de-sync prevention should be explored. As multiplayer is a bigger focus in later versions, this will not be address now.