Game Events Created
Created global game event script for handling events that may effect multiple scripts like when exiting the game or a menu. The GameEventManager singleton can be used to trigger or subscribe to any events that have been added to the dictionary. These will be added via the inspector to ease the process of game design.