Corrected Menu State
After reviewing my code, I found deprecated method and logic involving the menu player state. I was able to simplify this logic and remove unneeded method with logic in player state manager.

This change will help avoid confusion in the future as it removes the need for UI documents be loaded as part of the player menu state.